Split out mod.rs
